Back-end set up with liveblocks
I created functions with rooms that are dynamic to each user so multiple people can access a document and you store all the info on the users regarding that document. I can also create a new document, but won't show the editor yet. However, it keeps track of the document and associates it with the user.