App Complete - Sharing Fixed, Notifications, Collaboration
The user can now add and remove collaborators from the documents. They can also set the collaborator as a "viewer" or "editor". Upon mentioning a user in comments, the user that was mentioned will get a notification. If a user is added to the document, they will also get a notification. One problem for future fixes, you cannot share the document with someone who isn't in the database, or it will throw an error. They must register first with the email, then the document can be shared to them.